requirements.txt
pyyaml>=6.0
requests>=2.31.0
python-dateutil>=2.8.2
google-generativeai>=0.3.0